+++ packages/contracts/CHANGELOG.md
@@ -9,17 +9,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added support for release metadata.
+- Added `IMembership` and `IMembershipContract` and the `MembershipContractAnnounced` event.
 - Added a `metadata.json` file for all plugins.
 - Added `startDate` and `endDate` to the `ProposalCreated` event.
 - Added the `ProposalBase`, `Proposal`, and `ProposalUpgradeable` contracts to refactor and unify the proposal creation and execution process across all governance plugins.
 - Adds `startDate` and `endDate` to Multisig proposals. Execution and approvals of proposals have now to be in the boundaries of these 2 dates.
 - Adds `IEIP4824`
 - Adds `event NewURI` implementations for `daoURI()` and `setDaoURI()` definitions in `DAO`
+- Added check that plugin and DAO subdomains only contain `a-z`, `0-9` or a dash (`-`).
 - adds `allowFailureMap` to the plugins for creating proposals.
 
 ### Changed
 
 - Changed `CallbackReceived` which now emits the sender as well.
+- Renamed the `AddressesAdded` and `AddressesRemoved` event to `MembersAdded` and `MembersRemoved`.
 - Avoided integer division in the `MajorityVoting` execution criteria to avoid precision loss.
 - Replaced `ProposalParameters.minParticipation` by `minVotingPower` in `TokenVoting` and `AddresslistVoting`.
 - Rescaled and renamed `PCT_BASE = 10**18` to `RATIO_BASE = 10**6`.
@@ -30,6 +34,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed all occurences of `oracle` to `condition`.
 - Renamed `ItemSingleTarget` and `ItemMultiTarget` in `PermissionLib` to `SingleTargetPermission` and `MultiTargetPermission`.
 - Renamed `bulkOnSingleTarget()` and `bulkOnMultiTarget()` in `PermissionManager` to `applySingleTargetPermissions()` and `applyMultiTargetPermissions()`
+- Renames plugins `TokenVoting`, `AddresslistVoting`, `Admin` and `multisig` to be compliant to the new naming rules.
+- Renames `name` in `PluginRepoRegistry` and `DAORegistry` to `subdomain`.
 
 ### Removed